Migrate 10 `NetStru1` methods

+++ src/net.h
@@ -107,7 +107,7 @@ public:
 
 public:
     void FUN_0051ce86(uint32_t msg_type, uint32_t player_id, Player* recpt);
-    void FUN_0051ceac(uint32_t id, Player* recpt);
+    void FUN_0051ceac(uint8_t id, Player* recpt);
     void FUN_0051c748(Player* recpt);
     void sub_51C61E(Token* token); // Remove unit/sack from network tracking
     void FUN_0051d49b(Player* recpt);
